requirements.txt
aioredis==1.3.1
asgiref==3.9.1
async-timeout==5.0.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
cffi==2.0.0
channels==3.0.5
channels-redis==3.4.0
click==8.3.0
colorama==0.4.6
constantly==23.10.4
cryptography==46.0.1
daphne==3.0.2
Django==3.2.22
djangorestframework==3.12.4
h11==0.16.0
hiredis==3.2.1
huey==2.5.3
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
msgpack==1.1.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pyOpenSSL==25.3.0
pytz==2025.2
PyYAML==6.0.2
redis==6.4.0
service-identity==24.2.0
setuptools==80.9.0
sqlparse==0.5.3
Twisted==25.5.0
txaio==25.6.1
typing_extensions==4.15.0
tzdata==2025.2
uvicorn==0.36.0
whitenoise==6.11.0
zope.interface==8.0
